test: add a base test class for other ui tests
Add a base test which contains the initial setup of a test with Hilt, and with some flags to
initialize default data
