+++ app/src/androidTest/java/com/chesire/nekome/features/search/SearchTests.kt
@@ -1,26 +1,18 @@
 package com.chesire.nekome.features.search
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.chesire.nekome.ActivityTest
 import com.chesire.nekome.core.Resource
-import com.chesire.nekome.helpers.launchActivity
-import com.chesire.nekome.helpers.login
 import com.chesire.nekome.injection.DatabaseModule
 import com.chesire.nekome.injection.SearchModule
-import com.chesire.nekome.kitsu.AuthProvider
 import com.chesire.nekome.robots.activity
 import com.chesire.nekome.robots.search.search
 import com.chesire.nekome.search.api.SearchApi
 import dagger.hilt.android.testing.BindValue
-import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import dagger.hilt.android.testing.UninstallModules
 import io.mockk.coEvery
 import io.mockk.mockk
-import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
-import javax.inject.Inject
 
 private const val GENERIC_ERROR = "GENERIC_ERROR"
 private const val NO_RESULTS_ERROR = "NO_RESULTS_ERROR"
@@ -30,14 +22,7 @@ private const val NO_RESULTS_ERROR = "NO_RESULTS_ERROR"
     DatabaseModule::class,
     SearchModule::class
 )
-@RunWith(AndroidJUnit4::class)
-class SearchTests {
-
-    @get:Rule
-    val hilt = HiltAndroidRule(this)
-
-    @Inject
-    lateinit var authProvider: AuthProvider
+class SearchTests() : ActivityTest() {
 
     @BindValue
     val searchApi = mockk<SearchApi> {
@@ -51,12 +36,6 @@ class SearchTests {
         } coAnswers {
             Resource.Success(listOf())
         }
-    }
-
-    @Before
-    fun setUp() {
-        hilt.inject()
-        authProvider.login()
     }
 
     @Test
